fixed user integration for all, fixed sidebar for posts
