+++ src/main/resources/AccueilAdmin.fxml
@@ -18,7 +18,7 @@
                     </image>
                 </ImageView>
                 <Button mnemonicParsing="false" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" stylesheets="@css/Style.css" text="Accueil" />
-                <Button fx:id="Packages" mnemonicParsing="false" onAction="#SwitchToPackages" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" stylesheets="@css/Style.css" text="Voyages">
+                <Button fx:id="Voyages" mnemonicParsing="false" onAction="#SwitchToVoyages" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" stylesheets="@css/Style.css" text="Voyages">
                     <VBox.margin>
                         <Insets top="20.0" />
                     </VBox.margin>
