+++ src/main/resources/Posts.fxml
@@ -18,12 +18,12 @@
                     </image>
                 </ImageView>
                 <Button mnemonicParsing="false" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" text="Accueil" />
-                <Button mnemonicParsing="false" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" text="Packages">
+                <Button fx:id="Voyages" mnemonicParsing="false" onAction="#SwitchToVoyages" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" text="Voyages">
                     <VBox.margin>
                         <Insets top="20.0" />
                     </VBox.margin>
                 </Button>
-                <Button mnemonicParsing="false" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" text="Hotels">
+                <Button fx:id="Hotels" mnemonicParsing="false" onAction="#SwitchToHotels" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" text="Hotels">
                     <VBox.margin>
                         <Insets top="20.0" />
                     </VBox.margin>
@@ -38,12 +38,12 @@
                         <Insets top="20.0" />
                     </VBox.margin>
                 </Button>
-                <Button mnemonicParsing="false" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" text="Profile">
+                <Button fx:id="Profile" mnemonicParsing="false" onAction="#switchToProfile" prefHeight="50.0" prefWidth="230.0" styleClass="buttonNavigation" text="Profile">
                     <VBox.margin>
                         <Insets top="20.0" />
                     </VBox.margin>
                 </Button>
-                <Button mnemonicParsing="false" prefHeight="50.0" prefWidth="230.0" styleClass="buttonDeconnexion" text="Déconnexion">
+                <Button fx:id="Decon" mnemonicParsing="false" onAction="#deconnexion" prefHeight="50.0" prefWidth="230.0" styleClass="buttonDeconnexion" text="Déconnexion">
                     <VBox.margin>
                         <Insets top="20.0" />
                     </VBox.margin>
@@ -53,20 +53,16 @@
 
         <VBox layoutX="234.0" prefHeight="570.0" prefWidth="616.0" spacing="10.0">
             <padding>
-                <Insets top="10" right="10" bottom="10" left="10" />
+                <Insets bottom="10" left="10" right="10" top="10" />
             </padding>
             <HBox spacing="10.0">
-                <TextArea fx:id="postInput" prefHeight="50.0" prefWidth="540.0"
-                          promptText="What's on your mind?" HBox.hgrow="ALWAYS" />
-                <Button fx:id="postButton" text="Post" onAction="#handlePostButtonAction"
-                        prefHeight="50.0" prefWidth="100.0"
-                        style="-fx-background-color: #28a745; -fx-text-fill: white; -fx-font-size: 14px;
-                   -fx-border-radius: 5; -fx-background-radius: 5;" />
+                <TextArea fx:id="postInput" prefHeight="50.0" prefWidth="540.0" promptText="What's on your mind?" HBox.hgrow="ALWAYS" />
+                <Button fx:id="postButton" onAction="#handlePostButtonAction" prefHeight="50.0" prefWidth="100.0" style="-fx-background-color: #28a745; -fx-text-fill: white; -fx-font-size: 14px;                    -fx-border-radius: 5; -fx-background-radius: 5;" text="Post" />
             </HBox>
             <ScrollPane fitToWidth="true" VBox.vgrow="ALWAYS">
                 <VBox fx:id="postsContainer" spacing="10.0">
                     <padding>
-                        <Insets top="10" right="10" bottom="10" left="10" />
+                        <Insets bottom="10" left="10" right="10" top="10" />
                     </padding>
                 </VBox>
             </ScrollPane>
